requirements.txt
python-dotenv
freecurrencyapi